+++ cmd/vroom/flamegraph.go
@@ -169,3 +169,73 @@ func (env *environment) postFlamegraphFromChunksMetadata(w http.ResponseWriter, 
 	w.WriteHeader(http.StatusOK)
 	_, _ = w.Write(b)
 }
+
+type (
+	postFlamegraphBody struct {
+		Transaction []flamegraph.TransactionProfileCandidate `json:"transaction"`
+		Continuous  []flamegraph.ContinuousProfileCandidate  `json:"continuous"`
+	}
+)
+
+func (env *environment) postFlamegraph(w http.ResponseWriter, r *http.Request) {
+	ctx := r.Context()
+	hub := sentry.GetHubFromContext(ctx)
+	ps := httprouter.ParamsFromContext(ctx)
+	rawOrganizationID := ps.ByName("organization_id")
+	organizationID, err := strconv.ParseUint(rawOrganizationID, 10, 64)
+	if err != nil {
+		if hub != nil {
+			hub.CaptureException(err)
+		}
+		w.WriteHeader(http.StatusBadRequest)
+		return
+	}
+
+	hub.Scope().SetTag("organization_id", rawOrganizationID)
+
+	var body postFlamegraphBody
+	s := sentry.StartSpan(ctx, "processing")
+	s.Description = "Decoding data"
+	err = json.NewDecoder(r.Body).Decode(&body)
+	s.Finish()
+	if err != nil {
+		if hub != nil {
+			hub.CaptureException(err)
+		}
+		w.WriteHeader(http.StatusBadRequest)
+		return
+	}
+
+	s = sentry.StartSpan(ctx, "processing")
+	speedscope, err := flamegraph.GetFlamegraphFromCandidates(
+		ctx,
+		env.storage,
+		organizationID,
+		body.Transaction,
+		body.Continuous,
+		readJobs,
+	)
+	s.Finish()
+	if err != nil {
+		if hub != nil {
+			hub.CaptureException(err)
+		}
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	s = sentry.StartSpan(ctx, "json.marshal")
+	defer s.Finish()
+	b, err := json.Marshal(speedscope)
+	if err != nil {
+		if hub != nil {
+			hub.CaptureException(err)
+		}
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusOK)
+	_, _ = w.Write(b)
+}