+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Add endpoint to return a profile from a list of chunk IDs ([#470](https://github.com/getsentry/vroom/pull/470))
 - Add callTree generation for profile chunks ([#475](https://github.com/getsentry/vroom/pull/475))
 - Support generating a flamegraph from a list of chunk IDs ([#476](https://github.com/getsentry/vroom/pull/476))
+- Dual mode flamegraph endpoint ([#487](https://github.com/getsentry/vroom/pull/487))
 
 **Bug Fixes**: